feat: add source stubs, change freq type

+++ migrations/2025-10-07-143322-0000_create_log/up.sql
@@ -1,7 +1,7 @@
 -- Your SQL goes here
 CREATE TABLE logs (
     id SERIAL PRIMARY KEY,
-    frequency INTEGER NOT NULL,
+    frequency FLOAT(16) NOT NULL,
     xcoord  FLOAT(8) NOT NULL DEFAULT 0.0,
     ycoord FLOAT(8) NOT NULL DEFAULT 0.0,
     callsign VARCHAR(50) ,
